# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignableUserListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AssignableUserResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AssignableUserListResponse",
        "type": "object"
      },
      "AssignableUserResponse": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "ADMIN",
              "SOC_ANALYST"
            ],
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "display_name",
          "role"
        ],
        "title": "AssignableUserResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "PasswordResetRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "password",
          "expected_version"
        ],
        "title": "PasswordResetRequest",
        "type": "object"
      },
...
      "Role": {
        "enum": [
          "ADMIN",
          "SOC_ANALYST",
          "OT_ENGINEER",
          "READ_ONLY"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "const": true,
            "default": true,
            "title": "Authenticated",
            "type": "boolean"
          },
          "csrf_token": {
            "pattern": "^[A-Za-z0-9_-]{43}$",
            "title": "Csrf Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "user",
          "expires_at",
          "csrf_token"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "UserCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "username": {
            "maxLength": 64,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "display_name",
          "role",
          "password"
        ],
        "title": "UserCreateRequest",
        "type": "object"
      },
      "UserListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "UserListResponse",
        "type": "object"
      },
      "UserMutationResponse": {
        "additionalProperties": false,
        "properties": {
          "operation": {
            "enum": [
              "created",
              "updated",
              "password_reset"
            ],
            "title": "Operation",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "user",
          "operation"
        ],
        "title": "UserMutationResponse",
        "type": "object"
      },
      "UserPatchRequest": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "UserPatchRequest",
        "type": "object"
      },
      "UserResponse": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "password_changed_at": {
            "format": "date-time",
            "title": "Password Changed At",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "username",
          "display_name",
          "role",
          "active",
          "version",
          "created_at",
          "updated_at",
          "password_changed_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OTSOCSessionCookie": {
        "description": "Opaque HttpOnly local session cookie. The configured runtime name may differ.",
        "in": "cookie",
        "name": "otsoc_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Repository and contract foundation for the isolated, advisory-only OT-SOC Fusion X academic prototype.",
    "title": "OT-SOC Fusion X",
    "version": "1.1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Invalid local credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "read_session_api_v1_auth_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Read Session",
        "tags": [
          "authentication"
        ]
      }
    },
...
    "/api/v1/incident-assignees": {
      "get": {
        "operationId": "read_incident_assignees_api_v1_incident_assignees_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignableUserListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Read Incident Assignees",
        "tags": [
          "local-users"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "operationId": "read_users_api_v1_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Read Users",
        "tags": [
          "local-users"
        ]
      },
      "post": {
        "operationId": "create_user_api_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Create User",
        "tags": [
          "local-users"
        ]
      }
    },
    "/api/v1/users/{user_id}": {
      "patch": {
        "operationId": "patch_user_api_v1_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Patch User",
        "tags": [
          "local-users"
        ]
      }
    },
    "/api/v1/users/{user_id}/password-reset": {
      "post": {
        "operationId": "reset_user_password_api_v1_users__user_id__password_reset_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OTSOCSessionCookie": []
          }
        ],
        "summary": "Reset User Password",
        "tags": [
          "local-users"
        ]
      }
    },
...
  }
}